more fixes to fillGrid()
The existing code marked adjacents to the first tile of a pair as fillable before selecting the second, which is generally incorrect.  It also was too eager to mark a tile above the filled ones as being fillable (which should not be allowed directly if it is part of a lattice in that layer, and the lattice has already started to be filled).